removed logging.c
