+++ obc/sys/CMakeLists.txt
@@ -13,7 +13,6 @@ target_include_directories(OBC-firmware.out PUBLIC
 
 target_sources(OBC-firmware.out PUBLIC
     ${CMAKE_CURRENT_SOURCE_DIR}/fs_wrapper/obc_reliance_fs.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/logging/obc_logging.c
     ${CMAKE_CURRENT_SOURCE_DIR}/print/obc_print.c
     ${CMAKE_CURRENT_SOURCE_DIR}/persistent/obc_persistent.c
     ${CMAKE_CURRENT_SOURCE_DIR}/state/obc_state.c
